db-age-notifier/requirements.txt
influxdb
python-dotenv
requests
slack_sdk